#117: moving future/obsolete server code elsewhere, wanting to keep the current files simple
